Fixed PR issues
