+++ activerecord/lib/active_record/associations/singular_association.rb
@@ -5,8 +5,6 @@ module ActiveRecord
     class SingularAssociation < Association #:nodoc:
       # Implements the reader method, e.g. foo.bar for Foo.has_one :bar
       def reader
-        super
-
         if !loaded? || stale_target?
           reload
         end
